backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pgvector
psycopg2-binary
langchain
openai
python-multipart
